Change output of db iso date methods to use template literals
