+++ src/Query.ts
@@ -11,15 +11,19 @@ import type {
 	DictionaryTypeDefinitionNumber,
 	DictionaryTypeDefinitionString,
 	FlattenDocument,
-	ISOCalendarDate,
-	ISOCalendarDateTime,
-	ISOTime,
 	SchemaDefinition,
 	SchemaTypeDefinition,
 } from './Schema';
 import type Schema from './Schema';
 import type { SchemaTypeDefinitionScalar } from './schemaType';
-import type { DbDocument, DbSubroutineInputFind, DbSubroutineSetupOptions } from './types';
+import type {
+	DbDocument,
+	DbSubroutineInputFind,
+	DbSubroutineSetupOptions,
+	ISOCalendarDate,
+	ISOCalendarDateTime,
+	ISOTime,
+} from './types';
 
 // #region Types
 export interface QueryConstructorOptions<TSchema extends Schema | null> {
