+++ src/index.ts
@@ -17,9 +17,6 @@ export {
 	type SchemaDefinition,
 	type DictionariesOption,
 	type SchemaConstructorOptions,
-	type ISOCalendarDate,
-	type ISOCalendarDateTime,
-	type ISOTime,
 	type FlattenDocument,
 	type InferDocumentObject,
 	type InferModelObject,
@@ -55,4 +52,11 @@ export type {
 	SchemaTypeDefinitionNumber,
 	SchemaTypeDefinitionString,
 } from './schemaType';
-export type { EncryptFn, DecryptFn, MvRecord } from './types';
+export type {
+	EncryptFn,
+	DecryptFn,
+	MvRecord,
+	ISOCalendarDate,
+	ISOCalendarDateTime,
+	ISOTime,
+} from './types';
